Extract high level engine function
